requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Pillow==10.4.0
werkzeug==3.0.6
click==8.1.7
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
requests==2.32.3
email-validator==2.2.0
gunicorn==22.0.0
python-dotenv==1.0.1
detect-secrets==1.5.0
black==24.4.2
typing-extensions>=4.12.0
pytest==8.1.1
beautifulsoup4==4.12.3